avoid duplicated code because its prone to human error
